requirements.txt
azure-functions
appdirs==1.4.3
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
cfgv==3.0.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cymem==2.0.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
fuzzywuzzy==0.18.0
identify==1.4.11
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
murmurhash==1.0.2
nltk==3.4.5
nodeenv==1.3.5
numpy==1.18.1
packaging==20.1
pandas==1.0.1
plac==1.1.3
pluggy==0.13.1
pre-commit==2.1.0
preshed==3.0.2
py==1.8.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
sense2vec==1.0.2
six==1.14.0
spacy==2.2.3
srsly==1.0.1
thinc==7.3.1
toml==0.10.0
tox==3.14.5
tqdm==4.43.0
translate==3.5.0
urllib3==1.25.8
uuid==1.30
virtualenv==20.0.5
wasabi==0.6.0
Werkzeug==1.0.0
zipp==3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz